requirements.txt
Django==3.0.9
Pillow==7.2.0